Improve session termination
The session termination was hard coded to terminate Xfce4 sessions.
Add support to terminate unity (standard and "2d") and gnome sessions
(standard, classic and fallback). This uses the "DESKTOP_SESSION"
environment. If the session type is unrecognised, print an error in a
dialog, this isn't ideal but hopefully anyone using this will test it
first...

signed-off-by: Martyn Welch <martyn@welchs.me.uk>